backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# BigQuery
google-cloud-bigquery==3.18.0
google-cloud-bigquery-storage==2.23.0
db-dtypes==1.2.0

# Redis & Cache
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.2.1
structlog>=24.1.0

# Tasks
celery==5.3.6

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.20.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Reports & Documents
python-docx==1.1.0
jinja2==3.1.3
reportlab==4.1.0
pandas==2.2.0
openpyxl==3.1.2

# Causal Inference Engine (Módulo 5 — Impacto Econômico)
statsmodels==0.14.1
scipy==1.12.0
numpy>=1.26.0
scikit-learn==1.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# Development
black==24.1.1
ruff==0.1.9
mypy==1.8.0
pre-commit==3.6.0
